:white_check_mark: Write article.test.js
